display: Add simple RGB color balance transform
This adds support for adjusting the display's RGB color balance
globally via ColorDisplayManager. This is implemented as a simple
color transformation matrix that scales each channel.

The effect is similar to LineageOS' LiveDisplay feature, but being a
native color transform, it doesn't require changes to native code,
SELinux policies, or custom HALs.

Change-Id: Ia7ce363e3042ecfae438e452dbf155811a6bedeb
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>